fix: some fixes to the MerkleProof serialization
